a2a/pyproject.toml
[project]
name = "adk-x402-demo"
version = "0.1.0"
description = "ADK x402 Payment Protocol Demo using refactored x402_a2a package"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk",
    "google-adk>=0.5.0",
    "click>=8.2.0",
    "httpx>=0.28.1",
    "uvicorn>=0.34.2",
    "web3>=7.10.0",
    "cdp-sdk>=1.20.0",
    "x402-a2a",
    "bankofai-x402[tron]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["server", "client_agent"]

[project.scripts]
server = "server.__main__:main"



[tool.uv.sources]
x402-a2a = { git = "https://github.com/leo8tron/a2a-x402", branch = "feat/x402_tron", subdirectory = "python/x402_a2a" }
bankofai-x402 = { git = "https://github.com/bankofai/x402", tag = "v0.3.1", subdirectory = "python/x402" }

[tool.ruff]
target-version = "py313"

[dependency-groups]
lint = [
    "ruff>=0.13.1",
]
